Created code davinci prompts
Also organised and fixed the test cases
