+++ LLM Testing/ChatGPT-API/response.json
@@ -1,4 +1,5 @@
-{  "id": "chatcmpl-7RFPhqNV095y0CpnNpK9LoGNzeqiO",
+{
+  "id": "chatcmpl-7RFPhqNV095y0CpnNpK9LoGNzeqiO",
   "object": "chat.completion",
   "created": 1686728461,
   "model": "gpt-3.5-turbo-0301",
